made composer class and markov subclass, updated README and examples
Added the Markov class with all the functionality of the first
scripts and a bit more, and removed those scripts. Also included
a base class "Composer" that is useless as of now but may become
helpful when more composers are added, as I find things that can
be abstracted away into it.

Updated the README and the example file to reflect the new
changes, and improved wording a bit.